updated search-bar
