+++ src/style.css
@@ -21,8 +21,9 @@ body {
   align-items: center;
   text-align: center;
   gap: 20px;
-  margin: 100px;
-  margin-left: 15%;
+  margin-top: 30px;
+  margin-left: auto;
+  margin-right: auto;
 }
 .pagination {
   grid-column: 1 / -1;
@@ -40,7 +41,7 @@ body {
   display: flex;
   justify-content: center;
   align-items: center;
-  width: 30%;
+  width: 33%;
   margin: auto;
   min-width: 200px;
   background-color: white;
